Making gen_env_or_default available for the crate
